item_based

+++ distances.h
@@ -10,13 +10,13 @@
 
 using namespace std;
 
-auto distance_comparer = [](const std::pair<double, size_t> &a, const std::pair<double, size_t> &b) {
+auto distance_comparer = [](const std::pair<float, size_t> &a, const std::pair<float, size_t> &b) {
     return (a.first < b.first && a.second != b.second);
 };
 
 
-double cosine_distance(const double *vector1, const double *vector2, size_t size_cols);
+float cosine_distance(const float *vector1, const float *vector2, size_t size_cols);
 
-double manhattan_distance(vector<double> vector1, vector<double> vector2);
+float manhattan_distance(vector<float> vector1, vector<float> vector2);
 
 #endif //TP1_RECSYS_DISTANCES_H
